Split GameListing controller into LobbyWatcherController
- Splits the GameListingClient into two, helps comply with interface
segregation principle. The game update actions: post, update, remove,
are only done by lobby watcher, while fetch games and boot game are
done by a player from the lobby the two sets of interfaces are used
in a mutually exclusive way)
