+++ http-clients/src/main/java/org/triplea/http/client/lobby/game/listing/GameListingFeignClient.java
@@ -12,18 +12,6 @@ interface GameListingFeignClient {
   @RequestLine("GET " + GameListingClient.FETCH_GAMES_PATH)
   List<LobbyGameListing> fetchGameListing(@HeaderMap Map<String, Object> headers);
 
-  @RequestLine("POST " + GameListingClient.POST_GAME_PATH)
-  String postGame(@HeaderMap Map<String, Object> headers, LobbyGame lobbyGame);
-
-  @RequestLine("POST " + GameListingClient.UPDATE_GAME_PATH)
-  void updateGame(@HeaderMap Map<String, Object> headers, UpdateGameRequest updateGameRequest);
-
-  @RequestLine("POST " + GameListingClient.REMOVE_GAME_PATH)
-  void removeGame(@HeaderMap Map<String, Object> headers, String gameId);
-
-  @RequestLine("POST " + GameListingClient.KEEP_ALIVE_PATH)
-  boolean sendKeepAlive(@HeaderMap Map<String, Object> headers, String gameId);
-
   @RequestLine("POST " + GameListingClient.BOOT_GAME_PATH)
   void bootGame(@HeaderMap Map<String, Object> headers, String gameId);
 }
