+++ http-server/src/main/java/org/triplea/server/lobby/game/listing/LobbyWatcherControllerFactory.java
@@ -1,0 +1,11 @@
+package org.triplea.server.lobby.game.listing;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class LobbyWatcherControllerFactory {
+  public static LobbyWatcherController buildController(final GameListing gameListing) {
+    return LobbyWatcherController.builder().gameListing(gameListing).build();
+  }
+}
